Increment 3 Features
Added spotlight search. Enemy has a light that shines on vision cone. If the enemy is alerted it turns red if the enemy is patroling it is white. I also modified the damage modifier and revamped the enemies movement system. Before the enemy would go through walls and not be able to go around obstacles when patrolling. Fixed many of the movement mechanics of the enemy.